requirements.txt
streamlit==1.29.0
groq==0.4.1
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0
pydantic==2.5.2
cryptography>=41.0.0